add owner id condition in delete

+++ internal/handlers/listings.go
@@ -4,7 +4,6 @@ import (
 	"database/sql"
 	"encoding/json"
 	"errors"
-	"fmt"
 	"log"
 	"log/slog"
 	"net/http"
@@ -87,12 +86,19 @@ func (lh ListingHandler) Delete(w http.ResponseWriter, r *http.Request) {
 	requestId := middleware.RequestIdFromContext(ctx)
 	id := r.PathValue("id")
 
+	userID, ok := middleware.UserIDFromContext(ctx)
+	if !ok {
+		lh.logger.Error("no userid found in context")
+		httpx.Error(w, http.StatusInternalServerError, "something went wrong", httpx.CodeInternalError)
+		return
+	}
+
 	// lh.logger.Debug("debug log", "listing_id", id)
 	// lh.logger.Info("starting query", "listing_id", id)
 	// lh.logger.Warn("warn log", "listing_id", id)
 
 	_, err := lh.db.ExecContext(ctx,
-		`DELETE FROM listings WHERE id=$1`, id)
+		`DELETE FROM listings WHERE id = $1 AND user_id = $2`, id, userID)
 	if err != nil {
 		lh.logger.Error("delete failed", "listing_id", id, "requestId", requestId, "err", err)
 		// http.Error(w, "internal server error", http.StatusInternalServerError)
@@ -108,7 +114,6 @@ func (lh ListingHandler) Create(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	requestId := middleware.RequestIdFromContext(ctx)
 	userID, ok := middleware.UserIDFromContext(ctx)
-	fmt.Println("userID ==> ", userID)
 	if !ok {
 		lh.logger.Error("no userid found in context")
 		httpx.Error(w, http.StatusInternalServerError, "something went wrong", httpx.CodeInternalError)
